# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MeetMind Meeting Calendar API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Meeting {
  id: number;
  title: string;
  description?: string | null;
  startTime: string;
  endTime?: string | null;
  timezone?: string | null;
  location?: string | null;
  organizer?: string | null;
  meetingUrl?: string | null;
  notes?: string | null;
  reminderMinutes?: number | null;
  color?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateMeetingBody {
  title: string;
  description?: string | null;
  startTime: string;
  endTime?: string | null;
  timezone?: string | null;
  location?: string | null;
  organizer?: string | null;
  meetingUrl?: string | null;
  notes?: string | null;
  reminderMinutes?: number | null;
  color?: string | null;
}

export interface UpdateMeetingBody {
  title?: string;
  description?: string | null;
  startTime?: string;
  endTime?: string | null;
  timezone?: string | null;
  location?: string | null;
  organizer?: string | null;
  meetingUrl?: string | null;
  notes?: string | null;
  reminderMinutes?: number | null;
  color?: string | null;
}

export interface DeleteResult {
  success: boolean;
}

export interface ExtractMeetingBody {
  /** Base64 encoded image data */
  imageBase64: string;
  /** MIME type of the image (e.g. image/jpeg, image/png) */
  mimeType: string;
}

export interface ExtractedMeeting {
  title?: string | null;
  description?: string | null;
  startTime?: string | null;
  endTime?: string | null;
  timezone?: string | null;
  location?: string | null;
  organizer?: string | null;
  meetingUrl?: string | null;
  notes?: string | null;
  confidence?: number | null;
}