# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MeetMind Meeting Calendar API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all meetings
 */
export const GetMeetingsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  startTime: zod.date(),
  endTime: zod.date().nullish(),
  timezone: zod.string().nullish(),
  location: zod.string().nullish(),
  organizer: zod.string().nullish(),
  meetingUrl: zod.string().nullish(),
  notes: zod.string().nullish(),
  reminderMinutes: zod.number().nullish(),
  color: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const GetMeetingsResponse = zod.array(GetMeetingsResponseItem);

/**
 * @summary Create a new meeting
 */
export const CreateMeetingBody = zod.object({
  title: zod.string(),
  description: zod.string().nullish(),
  startTime: zod.date(),
  endTime: zod.date().nullish(),
  timezone: zod.string().nullish(),
  location: zod.string().nullish(),
  organizer: zod.string().nullish(),
  meetingUrl: zod.string().nullish(),
  notes: zod.string().nullish(),
  reminderMinutes: zod.number().nullish(),
  color: zod.string().nullish(),
});

/**
 * @summary Get a meeting by ID
 */
export const GetMeetingParams = zod.object({
  id: zod.coerce.number(),
});

export const GetMeetingResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  startTime: zod.date(),
  endTime: zod.date().nullish(),
  timezone: zod.string().nullish(),
  location: zod.string().nullish(),
  organizer: zod.string().nullish(),
  meetingUrl: zod.string().nullish(),
  notes: zod.string().nullish(),
  reminderMinutes: zod.number().nullish(),
  color: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a meeting
 */
export const UpdateMeetingParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMeetingBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().nullish(),
  startTime: zod.date().optional(),
  endTime: zod.date().nullish(),
  timezone: zod.string().nullish(),
  location: zod.string().nullish(),
  organizer: zod.string().nullish(),
  meetingUrl: zod.string().nullish(),
  notes: zod.string().nullish(),
  reminderMinutes: zod.number().nullish(),
  color: zod.string().nullish(),
});

export const UpdateMeetingResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  startTime: zod.date(),
  endTime: zod.date().nullish(),
  timezone: zod.string().nullish(),
  location: zod.string().nullish(),
  organizer: zod.string().nullish(),
  meetingUrl: zod.string().nullish(),
  notes: zod.string().nullish(),
  reminderMinutes: zod.number().nullish(),
  color: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a meeting
 */
export const DeleteMeetingParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteMeetingResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Extract meeting details from an image using AI
 */
export const ExtractMeetingFromImageBody = zod.object({
  imageBase64: zod.string().describe("Base64 encoded image data"),
  mimeType: zod
    .string()
    .describe("MIME type of the image (e.g. image\/jpeg, image\/png)"),
});

export const ExtractMeetingFromImageResponse = zod.object({
  title: zod.string().nullish(),
  description: zod.string().nullish(),
  startTime: zod.string().nullish(),
  endTime: zod.string().nullish(),
  timezone: zod.string().nullish(),
  location: zod.string().nullish(),
  organizer: zod.string().nullish(),
  meetingUrl: zod.string().nullish(),
  notes: zod.string().nullish(),
  confidence: zod.number().nullish(),
});