# lib/api-zod/src/generated/types/extractMeetingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MeetMind Meeting Calendar API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractMeetingBody {
  /** Base64 encoded image data */
  imageBase64: string;
  /** MIME type of the image (e.g. image/jpeg, image/png) */
  mimeType: string;
}