# lib/api-zod/src/generated/types/extractedMeeting.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MeetMind Meeting Calendar API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedMeeting {
  title?: string | null;
  description?: string | null;
  startTime?: string | null;
  endTime?: string | null;
  timezone?: string | null;
  location?: string | null;
  organizer?: string | null;
  meetingUrl?: string | null;
  notes?: string | null;
  confidence?: number | null;
}