# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MeetMind Meeting Calendar API
 * OpenAPI spec version: 0.1.0
 */

export interface Meeting {
  id: number;
  title: string;
  description?: string | null;
  startTime: Date;
  endTime?: Date | null;
  timezone?: string | null;
  location?: string | null;
  organizer?: string | null;
  meetingUrl?: string | null;
  notes?: string | null;
  reminderMinutes?: number | null;
  color?: string | null;
  createdAt: Date;
  updatedAt: Date;
}